backend/requirements.txt
django==5.1.6
graphene-django==3.2.2
pymongo==4.10.1
django-cors-headers==4.6.0
PyJWT==2.10.1
gunicorn==23.0.0
